Remove check for ACI and ClusterInstallRef when reconciling ClusterDeployment
This was preventing the creation of the agent cluster install when the
ClusterDeployment was created with the ClusterInstallRef already in
place.

Additionally the logic for `agentClusterInstallExists` was incorrect (it
should have returned true only when the error was nil and probably also
should have been checking the error type specifically for not found).
But also there is no reason to not ensure the ACI is correct even when
it already exists - `ensureAgentClusterInstall` should be idempotent.