Deploy providers in the same namespace
To do this it is required to remove the namespace and object prefix
fields from the default kustomization.yaml. Without these changes the
namespace objects in the manager manifest would get prefixed and thus
would change between the providers.

Now it will be required to directly name and namespace resources
when creating them.